Fix: move getMostRelatedPlace() to place service
