+++ src/places/places.module.ts
@@ -6,6 +6,7 @@ import { Place, PlaceSchema } from 'src/schemas/place.schema';
 import { Accommodation, AccommodationSchema } from 'src/schemas/accommodation.schema';
 import { Attraction, AttractionSchema } from 'src/schemas/attraction.schema';
 import { Restaurant, RestaurantSchema } from 'src/schemas/restaurant.schema';
+import { TagsModule } from 'src/tags/tags.module';
 
 @Module({
   imports: [MongooseModule.forFeatureAsync([
@@ -18,7 +19,8 @@ import { Restaurant, RestaurantSchema } from 'src/schemas/restaurant.schema';
       schema.discriminator(Restaurant.name, RestaurantSchema);
       return schema;
     }
-  }])],
+  }]),
+    TagsModule],
   controllers: [PlacesController],
   providers: [PlacesService],
   exports: [MongooseModule, PlacesService],
